Updated url encoding for fetch_waterbody and readme file
This commit fixes a bug in the fetch_waterbody function that was
preventing the query strings from being encoded completely.  It also
updates the readme file by adding the ggplot:: prefix the aes function
calls and removing the duplicate field name from the print statements
in the waterbody example.  Everything seems to work as it should.